Refactor rectangle to create a square

+++ src/main/com/tw/step/devbootcamp/shapes/TwoDimensionalShape.java
@@ -1,6 +1,6 @@
 package com.tw.step.devbootcamp.shapes;
 
 public interface TwoDimensionalShape {
-	public int area();
-	public int perimeter();
+	public double area();
+	public double perimeter();
 }
